put profile tests in a separate file
